Question 1: Done (construct drawing at the end -- change name of test.c file) Question 2: Done (improve Error Reporting at the end) Question 3: Added extra features: displayImage, setColourGradient and fillOval

+++ src/drawapptest/ImagePanel.java
@@ -80,14 +80,14 @@ public class ImagePanel
     }
     
     public void drawImage(String pathWay, int x, int y, int width, int height)
-    {
+    {  
        Image image = new Image(pathWay);
-       gc.drawImage(image, x, y, width, height);
+       gc.drawImage(image,x,y,width,height);
       
     }
     
-    public void fillOval(int x, int y, int width, int height)
+    public void fillOval(double x, double y, double width, double height)
     {
-        gc.fillOval(x, y, width, height);
+       gc.fillOval(x, y, width, height);
     }
 }
